+++ lib/Dialect/Comb/CMakeLists.txt
@@ -17,7 +17,7 @@ add_circt_dialect_library(CIRCTComb
 
   LINK_LIBS PUBLIC
   CIRCTHW
-  CIRCTSupport
+  CIRCTSVExternal
   MLIRIR
   MLIRInferTypeOpInterface
    )
