+++ lib/Dialect/HW/CMakeLists.txt
@@ -21,7 +21,7 @@ add_circt_dialect_library(
   Support
   
   LINK_LIBS PUBLIC
-  CIRCTSupport
+  CIRCTSVExternal
   MLIRIR
 )
