+++ lib/Dialect/SV/External/CMakeLists.txt
@@ -1,0 +1,8 @@
+add_circt_library(CIRCTSVExternal
+  SVAttributes.cpp
+
+  DEPENDS
+
+  LINK_LIBS PUBLIC
+  MLIRIR
+)
