+++ lib/Dialect/Seq/Transforms/CMakeLists.txt
@@ -9,6 +9,7 @@ add_circt_dialect_library(CIRCTSeqTransforms
   CIRCTSeq
   CIRCTSupport
   CIRCTSV
+  CIRCTSVExternal
   MLIRIR
   MLIRPass
   MLIRTransformUtils
